backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
flask-jwt-extended==4.4.4
flask-sqlalchemy==3.0.3
pymysql==1.0.3
werkzeug==2.2.3
python-dotenv==1.0.0
bcrypt==4.0.1
marshmallow==3.19.0 